Move animation separation into own file
