+++ src/separate-animations.ts
@@ -1,11 +1,6 @@
-import { Accessor, Animation, Document, NodeIO, Property, PropertyType } from '@gltf-transform/core';
+import { Accessor, Animation, Document, Property, PropertyType } from '@gltf-transform/core';
 import { Extension, WriterContext } from '@gltf-transform/core';
-import { copyToDocument, createDefaultPropertyResolver, prune, quantize, reorder, resample } from '@gltf-transform/functions';
-import { ALL_EXTENSIONS, EXTMeshoptCompression } from '@gltf-transform/extensions';
-import fs from 'fs';
-import path from 'path';
-import { CONFIG, AnimationSeparatorConfig } from './config.js';
-import { MeshoptDecoder, MeshoptEncoder } from 'meshoptimizer';
+import { copyToDocument, createDefaultPropertyResolver } from '@gltf-transform/functions';
 
 export class AnimationTargetPatcher extends Extension {
 	static EXTENSION_NAME = 'AnimationTargetPatcher';
@@ -46,7 +41,7 @@ export class AnimationTargetPatcher extends Extension {
 	}
 }
 
-class AnimationSeparator {
+export class AnimationSeparator {
     private sourceDoc: Document;
     private animationMap: Record<string, string | string[]>;
     private targetCache = new Map<string, Map<number, number>>();
@@ -213,85 +208,3 @@ class AnimationSeparator {
         return documents;
     }
 }
-
-async function writeDoc(io: NodeIO, document: Document, outputPath: string, fileName: string, config: AnimationSeparatorConfig) {
-    let joinedOutputPath = outputPath;
-    if (config.outputSeparateFolders) {
-        joinedOutputPath = path.join(joinedOutputPath, fileName);
-    }
-    fs.mkdirSync(joinedOutputPath, { recursive: true });
-    joinedOutputPath = path.join(joinedOutputPath, `${fileName}.${config.outputGlb ? 'glb' : 'gltf'}`);
-    
-    // Optimization steps
-    await document.transform(
-        resample({
-            cleanup: false,
-        }),
-        reorder({
-            encoder: MeshoptEncoder, 
-            target: 'size',
-            cleanup: false,
-        }),
-        quantize({
-            quantizationVolume: 'mesh',
-            quantizePosition: 14,
-            quantizeNormal: 8, // 10,
-            quantizeTexcoord: 12,
-            quantizeColor: 8,
-            quantizeWeight: 8,
-            quantizeGeneric: 12,
-            normalizeWeights: true,
-            cleanup: false
-        }),
-        prune({
-            propertyTypes: [PropertyType.ACCESSOR],
-        })
-    );
-
-    document
-        .createExtension(EXTMeshoptCompression)
-        .setRequired(true)
-        .setEncoderOptions({
-            method: EXTMeshoptCompression.EncoderMethod.FILTER,
-        });
-
-    await io.write(joinedOutputPath, document);
-}
-
-async function transformGltf(inputPath: string, outputPath: string, chunkMap: Record<string, string | string[]>, config: AnimationSeparatorConfig) {
-    await MeshoptDecoder.ready;
-    await MeshoptEncoder.ready;
-
-    const io = new NodeIO()
-        .registerExtensions([...ALL_EXTENSIONS, AnimationTargetPatcher])
-        .registerDependencies({ 'meshopt.decoder': MeshoptDecoder })
-        .registerDependencies({ 'meshopt.encoder': MeshoptEncoder });
-
-
-    const srcDoc = await io.read(inputPath);
-    
-    const separator = new AnimationSeparator(srcDoc, chunkMap);
-    const { baseDoc, chunkDocs } = await separator.separate();
-
-    // Write the base document (model without extracted animations)
-    await writeDoc(io, baseDoc, outputPath, path.basename(inputPath, path.extname(inputPath)), config);
-
-    // Write each animation chunk
-    for (const [name, doc] of chunkDocs) {
-        await writeDoc(io, doc, outputPath, name, config);
-    }
-}
-
-
-async function main() {
-    try {
-        console.log(`Transforming GLTF...\n`);
-        await transformGltf(CONFIG.inputFile, CONFIG.outputPath, CONFIG.animationMap, CONFIG);
-        console.log("\nGLTF transformation complete.");
-    } catch (error) {
-        console.error("An error occurred during the transformation process:", error);
-        process.exit(1);
-    }
-}
-
-main();
